Fixed tracking error due to redo stack not cleared on external undo records added
Fixed:
When adding a record externally of the system that would NOT erase the redo-stack (selection change) while a custom undo record was in the redo-stack, that record would be incorrectly shifted with the undo stack.
When adding a record externally of the system that would erase the redo-stack (normal records) while a custom undo record was in the redo-stack, that record would not be erased correctly.