Fixed warning and playmode tracking bug
Removed warning by making collections leveled aswell (limited depth).
In order for them to be useful (anonymous functions can store a list) I had to create another level (3-level serialization). Have not found a way to do that without code duplication. It's ugly, but either that or the warning due to the loop...

Fixed playmode tracking bug when after exiting playmode, some 'Selection Changed' records would not be picked up and create a shift in tracking.